Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,8 +2,8 @@
 
 int main() 
 {
-    int r;
-    scanf("%d",&r);
+    float r;
+    scanf("%f",&r);
     printf("Area: %.2f",(3.14*r*r) );
     return 0;
 }
